# backend/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Mimi Platform API
  description: |
    REST API платформы Mimi — проектный менеджер с задачами, документами, канвасами и командами.

    **Аутентификация**: большинство эндпоинтов требуют `Authorization: Bearer <access_token>`.
    Получите токен через `POST /api/auth/login`.

    **Архитектура**: 9 независимых микросервисов, доступных через NGINX gateway на порту 8080.
  version: "1.0.0"
  contact:
    name: Mimi Dev Team
servers:
  - url: http://localhost:8080
    description: Локальный Docker Compose
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:    { type: string, example: Иван Иванов }
        email:   { type: string, format: email, example: ivan@example.com }
        password: { type: string, minLength: 6, example: secret123 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string }
    TokenResponse:
      type: object
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        expiresIn:    { type: integer, description: Секунд до истечения access token }
        user:         { $ref: '#/components/schemas/User' }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    UpdateProfileRequest:
      type: object
      properties:
        name:      { type: string }
        email:     { type: string, format: email }
        avatarUrl: { type: string, format: uri }
    ChangePasswordRequest:
      type: object
      required: [current, next]
      properties:
        current: { type: string }
        next:    { type: string, minLength: 6 }
    NotificationPrefs:
      type: object
      properties:
        taskAssigned:      { type: boolean }
        taskStatusChanged: { type: boolean }
        teamInvited:       { type: boolean }
        mentionedInDoc:    { type: boolean }

    # ── Users ─────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:        { type: string, format: uuid }
        name:      { type: string }
        email:     { type: string, format: email }
        avatarUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        notificationPrefs: { $ref: '#/components/schemas/NotificationPrefs' }
...
    ErrorResponse:
      type: object
      properties:
        code:    { type: string, example: validation }
        message: { type: string, example: Описание ошибки }
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        "201":
          description: Успешная регистрация — возвращает пару токенов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "409":
          description: Email уже занят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example: { code: email_taken, message: Email уже занят }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        "401":
          description: Неверный email или пароль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление токенов по refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        "200":
          description: Новая пара токенов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Выход (инвалидация refresh token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        "204": { description: Успешный выход }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Профиль текущего пользователя
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Auth]
      summary: Обновление профиля (имя, email, аватар)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "409": { description: Email уже занят }
  /api/auth/me/password:
    post:
      tags: [Auth]
      summary: Смена пароля
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        "204": { description: Пароль изменён }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/auth/me/notifications:
    patch:
      tags: [Auth]
      summary: Настройки уведомлений текущего пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationPrefs' }
      responses:
        "204": { description: Настройки сохранены }
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ═══════════════════════════════════════════════════════════════════
  # USERS
  # ═══════════════════════════════════════════════════════════════════
...